Rename OnBoardingPage to use PascalCase filename
Renamed the OnBoardingPage component file from onBoardingPage.tsx to OnBoardingPage.tsx and updated imports to use the new filename. This change ensures consistency with React component naming conventions.
