+++ app/components/auth/onboarding/OnBoardingPage.tsx
